stock_predictor/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pandas==1.3.5
numpy==1.21.6
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.36
textblob==0.17.1
ta==0.10.2
gunicorn==21.2.0 